separate hypermedia approaches
